move `Header` into own module
